+++ changelog.md
@@ -29,6 +29,7 @@ Upcoming release
 - Behavioral data diffusion analysis
   - :warning: EZ parameter recovery gave incorrect result when error rate was 50%
   - :warning: Rows are now ordered to be case-insensitive
+  - :warning: If any data is missing from a trial, the whole trial is dropped
 - Various output fixes
   - :warning: In pivot tables, rows are now ordered to be case-insensitive
 - Various data import fixes